UI tests for EditPrivateShareFragment
