+++ owncloudApp/src/androidTest/java/com/owncloud/android/shares/domain/publicShares/CreatePublicShareTest.kt
@@ -17,27 +17,20 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.owncloud.android.shares.domain
+package com.owncloud.android.shares.domain.publicShares
 
 import android.accounts.Account
 import android.accounts.AccountManager
 import android.content.Context
 import android.content.Intent
 import android.os.Parcelable
-import android.widget.DatePicker
 import androidx.lifecycle.MutableLiveData
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.click
-import androidx.test.espresso.action.ViewActions.replaceText
-import androidx.test.espresso.action.ViewActions.typeText
 import androidx.test.espresso.assertion.ViewAssertions.doesNotExist
 import androidx.test.espresso.assertion.ViewAssertions.matches
-import androidx.test.espresso.contrib.PickerActions
-import androidx.test.espresso.matcher.ViewMatchers.isChecked
 import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
-import androidx.test.espresso.matcher.ViewMatchers.isNotChecked
-import androidx.test.espresso.matcher.ViewMatchers.withClassName
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.espresso.matcher.ViewMatchers.withText
 import androidx.test.platform.app.InstrumentationRegistry
@@ -51,13 +44,13 @@ import com.owncloud.android.lib.common.accounts.AccountUtils
 import com.owncloud.android.lib.common.operations.RemoteOperationResult
 import com.owncloud.android.lib.resources.status.CapabilityBooleanType
 import com.owncloud.android.lib.resources.status.OwnCloudVersion
+import com.owncloud.android.shares.domain.OCShare
 import com.owncloud.android.shares.presentation.ShareActivity
 import com.owncloud.android.shares.presentation.OCShareViewModel
 import com.owncloud.android.ui.activity.FileActivity
 import com.owncloud.android.utils.AccountsManager
 import com.owncloud.android.utils.TestUtil
 import com.owncloud.android.vo.Resource
-import org.hamcrest.Matchers
 import org.junit.AfterClass
 import org.junit.Before
 import org.junit.BeforeClass
@@ -71,11 +64,8 @@ import org.koin.dsl.module
 import org.mockito.Mockito.`when`
 import org.mockito.Mockito.mock
 import org.mockito.Mockito.spy
-import java.text.DateFormat
-import java.text.SimpleDateFormat
-import java.util.Calendar
 
-class EditPublicShareTest {
+class CreatePublicShareTest {
     @Rule
     @JvmField
     val activityRule = ActivityTestRule(
@@ -87,9 +77,8 @@ class EditPublicShareTest {
     private lateinit var file: OCFile
 
     private val publicShares = arrayListOf(
-        TestUtil.createPublicShare( // With no expiration date
+        TestUtil.createPublicShare(
             path = "/Photos/image.jpg",
-            expirationDate = 0L,
             isFolder = false,
             name = "image.jpg link",
             shareLink = "http://server:port/s/1"
@@ -97,29 +86,14 @@ class EditPublicShareTest {
         TestUtil.createPublicShare(
             path = "/Photos/image.jpg",
             isFolder = false,
-            name = "image.jpg updated link",
+            name = "image.jpg link (2)",
             shareLink = "http://server:port/s/2"
         ),
         TestUtil.createPublicShare(
             path = "/Photos/image.jpg",
             isFolder = false,
-            name = "image.jpg changed again link",
+            name = "image.jpg link (3)",
             shareLink = "http://server:port/s/3"
-        ),
-        TestUtil.createPublicShare( // With password but not expiration date
-            path = "/Photos/image.jpg",
-            expirationDate = 0L,
-            isFolder = false,
-            name = "image.jpg link",
-            shareLink = "http://server:port/s/1",
-            shareWith = "1"
-        ),
-        TestUtil.createPublicShare( // With expiration date but not password
-            path = "/Photos/image.jpg",
-            expirationDate = 2587896257,
-            isFolder = false,
-            name = "image.jpg link",
-            shareLink = "http://server:port/s/1"
         )
     )
 
@@ -215,205 +189,144 @@ class EditPublicShareTest {
     }
 
     @Test
-    fun editName() {
+    fun createPublicShareWithNoPublicSharesYet() {
         loadCapabilitiesSuccessfully()
+        loadSharesSuccessfully(arrayListOf())
 
-        val existingPublicShare = publicShares[0]
-        loadSharesSuccessfully(arrayListOf(existingPublicShare))
+        // Create share
+        onView(withId(R.id.addPublicLinkButton)).perform(click())
 
-        val updatedPublicShare = publicShares[1]
+        val newPublicShare = publicShares[0]
+        savePublicShare(newPublicShare)
 
-        // 1. Open dialog to edit an existing public share
-        onView(withId(R.id.editPublicLinkButton)).perform(click())
-
-        // 2. Update fields
-        onView(withId(R.id.shareViaLinkNameValue)).perform(replaceText(updatedPublicShare.name))
-
-        // 3. Edit public share with success
-        editPublicShare(updatedPublicShare, "", resource = Resource.success())
-
-        // 4. Share properly updated
+        // New share properly created
         sharesLiveData.postValue(
             Resource.success(
-                arrayListOf(updatedPublicShare)
+                arrayListOf(newPublicShare)
             )
         )
 
-        // 5. Check whether the dialog to create the public share has been properly closed
-        onView(withText(R.string.share_via_link_edit_title)).check(doesNotExist())
-        onView(withText(updatedPublicShare.name)).check(matches(isDisplayed()))
+        // Check whether the dialog to create the public share has been properly closed
+        onView(withText(R.string.share_via_link_create_title)).check(doesNotExist())
+        onView(withText(newPublicShare.name)).check(matches(isDisplayed()))
     }
 
     @Test
-    fun addPassword() {
+    fun createPublicShareWithAlreadyExistingShares() {
         loadCapabilitiesSuccessfully()
+        val existingPublicShares = publicShares.take(2) as ArrayList<OCShare>
+        loadSharesSuccessfully(
+            existingPublicShares
+        )
 
-        val existingPublicShare = publicShares[0]
-        loadSharesSuccessfully(arrayListOf(existingPublicShare))
+        onView(withId(R.id.addPublicLinkButton)).perform(click())
 
-        val password = "1234"
+        val newPublicShare = publicShares[2]
+        savePublicShare(newPublicShare)
 
-        // 1. Open dialog to edit an existing public share
-        onView(withId(R.id.editPublicLinkButton)).perform(click())
-
-        // 2. Enable password and type it
-        onView(withId(R.id.shareViaLinkPasswordSwitch)).perform(click())
-        onView(withId(R.id.shareViaLinkPasswordValue)).perform(typeText(password))
-
-        // 3. Edit public share with success
-        editPublicShare(existingPublicShare, "1234", resource = Resource.success())
-
-        // 4. Share properly updated
-        val updatedPublicShare = publicShares[3]
-
+        // New share properly created
         sharesLiveData.postValue(
             Resource.success(
-                arrayListOf(updatedPublicShare)
+                publicShares
             )
         )
 
-        // 5. Open dialog to check whether the password has been properly set
-        onView(withId(R.id.editPublicLinkButton)).perform(click())
-        onView(withId(R.id.shareViaLinkPasswordSwitch)).check(matches(isChecked()))
+        // Check whether the dialog to create the public share has been properly closed
+        onView(withText(R.string.share_via_link_create_title)).check(doesNotExist())
+        onView(withText(newPublicShare.name)).check(matches(isDisplayed()))
     }
 
     @Test
-    fun removePassword() {
+    fun createMultiplePublicShares() {
         loadCapabilitiesSuccessfully()
+        loadSharesSuccessfully(arrayListOf())
 
-        val existingPublicShare = publicShares[3]
-        loadSharesSuccessfully(arrayListOf(existingPublicShare))
+        /**
+         * 1st public share
+         */
+        onView(withId(R.id.addPublicLinkButton)).perform(click())
 
-        // 1. Open dialog to edit an existing public share
-        onView(withId(R.id.editPublicLinkButton)).perform(click())
+        val newPublicShare1 = publicShares[0]
+        savePublicShare(newPublicShare1)
 
-        // 2. Disable password
-        onView(withId(R.id.shareViaLinkPasswordSwitch)).perform(click())
-
-        // 3. Edit public share with success
-        editPublicShare(existingPublicShare, "", resource = Resource.success())
-
-        // 4. Share properly updated
-        val updatedPublicShare = publicShares[0]
-
+        // New share properly created
         sharesLiveData.postValue(
             Resource.success(
-                arrayListOf(updatedPublicShare)
+                arrayListOf(newPublicShare1)
             )
         )
 
-        // 5. Open dialog to check whether the password has been properly disabled
-        onView(withId(R.id.editPublicLinkButton)).perform(click())
-        onView(withId(R.id.shareViaLinkPasswordSwitch)).check(matches(isNotChecked()))
-    }
+        // Check whether the dialog to create the public share has been properly closed
+        onView(withText(R.string.share_via_link_create_title)).check(doesNotExist())
+        onView(withText(newPublicShare1.name)).check(matches(isDisplayed()))
 
-    @Test
-    fun addExpirationDate() {
-        loadCapabilitiesSuccessfully()
+        /**
+         * 2nd public share
+         */
+        onView(withId(R.id.addPublicLinkButton)).perform(click())
 
-        val existingPublicShare = publicShares[0]
-        loadSharesSuccessfully(arrayListOf(existingPublicShare))
+        val newPublicShare2 = publicShares[1]
+        savePublicShare(newPublicShare2)
 
-        val calendar = Calendar.getInstance()
-        calendar.add(Calendar.DAY_OF_YEAR, 1)
-        val formatter: DateFormat = SimpleDateFormat("MMM dd, yyyy");
-        val expirationDate = formatter.format(calendar.time);
-        val publicLinkExpirationDateInMillis = SimpleDateFormat.getDateInstance().parse(expirationDate).time
-
-        // 1. Open dialog to edit an existing public share
-        onView(withId(R.id.editPublicLinkButton)).perform(click())
-
-        // 2. Enable expiration date and set it
-        onView(withId(R.id.shareViaLinkExpirationSwitch)).perform(click())
-        onView(withClassName(Matchers.equalTo(DatePicker::class.java.name))).perform(
-            PickerActions.setDate(
-                calendar.get(Calendar.YEAR),
-                calendar.get(Calendar.MONTH) + 1, // January code is 0, so let's add 1
-                calendar.get(Calendar.DATE)
-            )
-        );
-        onView(withId(android.R.id.button1)).perform(click())
-
-        // 3. Edit public share with success
-        editPublicShare(existingPublicShare, "", publicLinkExpirationDateInMillis, Resource.success())
-
-        // 4. Share properly updated
-        val updatedPublicShare = publicShares[4]
-
+        // New share properly created
         sharesLiveData.postValue(
             Resource.success(
-                arrayListOf(updatedPublicShare)
+                publicShares.take(2)
             )
         )
 
-        // 5. Open dialog to check whether the expiration date is enabled
-        onView(withId(R.id.editPublicLinkButton)).perform(click())
-        onView(withId(R.id.shareViaLinkExpirationSwitch)).check(matches(isChecked()))
-    }
+        // Check whether the dialog to create the public share has been properly closed
+        onView(withText(R.string.share_via_link_create_title)).check(doesNotExist())
+        onView(withText(newPublicShare2.name)).check(matches(isDisplayed()))
 
-    @Test
-    fun removeExpirationDate() {
-        loadCapabilitiesSuccessfully()
+        /**
+         * 3rd public share
+         */
+        onView(withId(R.id.addPublicLinkButton)).perform(click())
 
-        val existingPublicShare = publicShares[4]
-        loadSharesSuccessfully(arrayListOf(existingPublicShare))
+        val newPublicShare3 = publicShares[2]
+        savePublicShare(newPublicShare3)
 
-        // 1. Open dialog to edit an existing public share
-        onView(withId(R.id.editPublicLinkButton)).perform(click())
-
-        // 2. Disable expiration date
-        onView(withId(R.id.shareViaLinkExpirationSwitch)).perform(click())
-
-        // 3. Edit public share with success
-        editPublicShare(existingPublicShare, "", resource = Resource.success())
-
-        // 4. Share properly updated
-        val updatedPublicShare = publicShares[0]
-
+        // New share properly created
         sharesLiveData.postValue(
             Resource.success(
-                arrayListOf(updatedPublicShare)
+                publicShares
             )
         )
 
-        // 5. Open dialog to check whether the expiration date is disabled
-        onView(withId(R.id.editPublicLinkButton)).perform(click())
-        onView(withId(R.id.shareViaLinkExpirationSwitch)).check(matches(isNotChecked()))
+        // Check whether the dialog to create the public share has been properly closed
+        onView(withText(R.string.share_via_link_create_title)).check(doesNotExist())
+        onView(withText(newPublicShare3.name)).check(matches(isDisplayed()))
     }
 
     @Test
-    fun editShareLoading() {
+    fun createShareLoading() {
         loadCapabilitiesSuccessfully()
+        loadSharesSuccessfully(arrayListOf())
 
-        val existingPublicShare = publicShares[3]
-        loadSharesSuccessfully(arrayListOf(existingPublicShare))
+        onView(withId(R.id.addPublicLinkButton)).perform(click())
 
-        onView(withId(R.id.editPublicLinkButton)).perform(click())
-
-        editPublicShare(existingPublicShare, resource = Resource.loading())
+        savePublicShare(publicShares[0], Resource.loading())
 
         onView(withText(R.string.common_loading)).check(matches(isDisplayed()))
     }
 
     @Test
-    fun editShareError() {
+    fun createShareError() {
         loadCapabilitiesSuccessfully()
+        loadSharesSuccessfully(arrayListOf())
 
-        val existingPublicShare = publicShares[0]
-        loadSharesSuccessfully(arrayListOf(existingPublicShare))
+        onView(withId(R.id.addPublicLinkButton)).perform(click())
 
-        onView(withId(R.id.editPublicLinkButton)).perform(click())
-
-        editPublicShare(
-            existingPublicShare,
-            password = "",
-            resource = Resource.error(
+        savePublicShare(
+            publicShares[0],
+            Resource.error(
                 RemoteOperationResult.ResultCode.FORBIDDEN,
                 exception = Exception("Error when retrieving shares")
             )
         )
 
-        onView(withText(R.string.update_link_file_error)).check(matches(isDisplayed()))
+        onView(withText(R.string.share_link_file_error)).check(matches(isDisplayed()))
     }
 
     private fun getOCFileForTesting(name: String = "default") = OCFile("/Photos").apply {
@@ -441,19 +354,14 @@ class EditPublicShareTest {
         sharesLiveData.postValue(Resource.success(shares))
     }
 
-    private fun editPublicShare(
-        share: OCShare,
-        password: String? = null,
-        publicLinkExpirationDateInMillis: Long = -1,
-        resource: Resource<Unit> = Resource.success() // Expected result when editing the share
-    ) {
+    private fun savePublicShare(newShare: OCShare, resource: Resource<Unit> = Resource.success()) {
         `when`(
-            ocShareViewModel.updatePublicShare(
+            ocShareViewModel.insertPublicShare(
+                file.remotePath,
                 1,
-                share.name!!,
-                password,
-                publicLinkExpirationDateInMillis,
-                1,
+                newShare.name!!,
+                "",
+                -1,
                 false
             )
         ).thenReturn(
